+++ src/Bore/SpacecookieClone/Search/Search.hs
@@ -126,16 +126,6 @@ computeSelectorScore keywords selector =
   in
     (highlightedSelector, likenessSum)
 
-weightFarProximity :: Float
-weightFarProximity = 0.1  -- Example: Penalize for far distances
-
-weightCloseProximity :: Float
-weightCloseProximity = 2  -- Example: Reward for close distances
-
--- | Max distance between keywords in ordered proximity bonus before a penalty.
-thresholdDistance :: Int
-thresholdDistance = 1
-
 -- FIXME: I think longer documents will benefit more or something? there's something like that i'm not thinking of here.
 -- | Calculate proximity score for different keywords based on their positions. The closer
 -- the keywords are to each other, the higher the score. A significant bonus is applied for